% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_delay_relative}
\alias{check_delay_relative}
\title{Delay Effect Matrix}
\usage{
check_delay_relative(delay_type, delaytime_list, S_matrix, S_matrix_delay)
}
\arguments{
\item{delay_type}{vector of reaction type taking on the values 0, 1, or 2}

\item{delaytime_list}{a list representing the delay time of each reaction}

\item{S_matrix}{the stoichiometric matrix at the initiation time}

\item{S_matrix_delay}{the stoichiometric matrix at the completion time}
}
\value{
A 2-row delay_effect_matrix
}
\description{
Delay Effect Matrix
}
