% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_pdf}
\alias{convert_pdf}
\title{Probability Density Functions Conversion}
\usage{
convert_pdf(a_vector)
}
\arguments{
\item{a_vector}{a vector}
}
\value{
PDF
}
\description{
Probability Density Functions Conversion
}
