% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{picksample}
\alias{picksample}
\title{Results Sampling}
\usage{
picksample(list_output, i = 1, t)
}
\arguments{
\item{list_output}{a outcome of a single simulation}

\item{i}{species i}

\item{t}{time t}
}
\value{
species number
}
\description{
Results Sampling
}
