% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_mean}
\alias{plot_mean}
\title{The Mean Value of Species i at Time t}
\usage{
plot_mean(result, i, t)
}
\arguments{
\item{result}{result of solve_DelaySSA}

\item{i}{species i}

\item{t}{time t}
}
\value{
mean value
}
\description{
The Mean Value of Species i at Time t
}
