% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_SSA_density}
\alias{plot_SSA_density}
\title{Density at Time t_pick}
\usage{
plot_SSA_density(result, t_pick)
}
\arguments{
\item{result}{result of solve_DelaySSA}

\item{t_pick}{time t_pick}
}
\value{
a picture showing the density at t_pick
}
\description{
Density at Time t_pick
}
