% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{propensity_n}
\alias{propensity_n}
\title{For Propensity Calculation}
\usage{
propensity_n(n, reactant_matrx)
}
\arguments{
\item{n}{species number}

\item{reactant_matrx}{species reactant matrix}
}
\value{
a vector used for calculating propensity function
}
\description{
For Propensity Calculation
}
