% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct.R
\name{simulate_reaction}
\alias{simulate_reaction}
\title{Gillespie Algorithm}
\usage{
simulate_reaction(tmax, n_initial, t_initial, S_matrix, fun_fr)
}
\arguments{
\item{tmax}{cutoff time}

\item{n_initial}{initial species number}

\item{t_initial}{initial time}

\item{S_matrix}{the stoichiometric matrix at the initiation time}

\item{fun_fr}{a function representing the propensity function}
}
\value{
the amount of a species and the corresponding time
}
\description{
A Stochastic Simulation Algorithm (SSA) without Delays Using Gillespie Method
}
