% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_rejection.R
\name{simulate_reaction_delay_rejection}
\alias{simulate_reaction_delay_rejection}
\title{Delay Rejection Method Algorithm}
\usage{
simulate_reaction_delay_rejection(
  tmax,
  n_initial,
  t_initial,
  S_matrix,
  S_matrix_delay,
  k,
  fun_fr,
  delay_type,
  delaytime_list,
  delay_effect_matrix,
  reactant_matrix_delay
)
}
\arguments{
\item{tmax}{cutoff time}

\item{n_initial}{initial species number}

\item{t_initial}{initial time}

\item{S_matrix}{the stoichiometric matrix at the initiation time}

\item{S_matrix_delay}{the stoichiometric matrix at the completion time}

\item{k}{a reaction rate vector}

\item{fun_fr}{a function representing the propensity function}

\item{delay_type}{the reaction type vector taking on the values 0, 1, or 2}

\item{delaytime_list}{a list representing the delay time of each reaction}

\item{delay_effect_matrix}{a matrix representing that reaction without delay affects reaction with delay}

\item{reactant_matrix_delay}{species reactant matrix in delay part}
}
\value{
the amount of a species and the corresponding time
}
\description{
A Stochastic Simulation Algorithm (SSA) with Delays Using Rejection Method
}
