% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextreaction.R
\name{simulate_reaction_nextreaction}
\alias{simulate_reaction_nextreaction}
\title{Next Reaction Method Algorithm}
\usage{
simulate_reaction_nextreaction(tmax, n_initial, t_initial, S_matrix, k, fun_fr)
}
\arguments{
\item{tmax}{cutoff time}

\item{n_initial}{initial species number}

\item{t_initial}{initial time}

\item{S_matrix}{the stoichiometric matrix at the initiation time}

\item{k}{a reaction rate vector}

\item{fun_fr}{a function representing the propensity function}
}
\value{
the amount of a species and the corresponding time
}
\description{
A Stochastic Simulation Algorithm (SSA) without Delays Using Next Reaction Method
}
