% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_SSA_mean}
\alias{plot_SSA_mean}
\title{Mean of Species over Time}
\usage{
plot_SSA_mean(result, t = seq(0, tmax, by = 1))
}
\arguments{
\item{result}{result of solve_DelaySSA}

\item{t}{time series}
}
\value{
a picture showing the mean through time
}
\description{
Mean of Species over Time
}
