% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulation_DelaySSA}
\alias{simulation_DelaySSA}
\title{A Stochastic Simulation Algorithm (SSA) with Delays}
\usage{
simulation_DelaySSA(
  algorithm = "DelayMNR",
  sample_size,
  tmax,
  n_initial,
  t_initial,
  S_matrix,
  S_matrix_delay = NULL,
  k,
  reactant_matrix,
  reactant_matrix_delay = NULL,
  delay_type = NULL,
  delaytime_list = NULL
)
}
\arguments{
\item{algorithm}{"DelayMNR" (default), "DelayReject", or "DelayDirect". If without delay, "Direct", "MNR" and "NR"  are recommended and the parameters of "S_matrix_dalay" "delay_type" and "delaytime_list" can be omitted.}

\item{sample_size}{the number of repetitions}

\item{tmax}{cutoff time}

\item{n_initial}{initial species number}

\item{t_initial}{initial time}

\item{S_matrix}{the stoichiometric matrix at the initiation time}

\item{S_matrix_delay}{the stoichiometric matrix at the completion time}

\item{k}{a reaction rate vector}

\item{reactant_matrix}{species reactant matrix}

\item{reactant_matrix_delay}{species reactant matrix at delay part}

\item{delay_type}{the reaction type vector taking on the values 0, 1, or 2}

\item{delaytime_list}{a list representing the delay time of each reaction}
}
\value{
A list contains sublists including the amount of a species and the corresponding time
}
\description{
DelaySSA implements a stochastic simulation algorithm (SSA) with delays in R through different algorithms.
}
