% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tau_element}
\alias{tau_element}
\title{Delay Time}
\usage{
tau_element(element)
}
\arguments{
\item{element}{an element in delaytime_list which could be a function or a fixed number}
}
\value{
delaytime which is a fixed number or a stochastic value governed by a function
}
\description{
Delay Time
}
